pyproject.toml
[project]
name = "verbosa"
version = "0.1.0"
description = "A small library that streamlines everyday data handling tasks like validating data types, enforcing integrity rules, and applying common data checks. It helps you write cleaner, more reliable code with less effort."
authors = [
    {name = "Pérez Lázaro Leonardo", email = "leonardo.perezlazaro@gmail.com"}
]
license = "LICENSE"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "boto3 (>=1.42.1,<2.0.0)",
    "awswrangler (>=3.14.0,<4.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
